Upgrade osmdroid; restrict max/min latitude (#246)
* Upgrade to latest osmdroid 6.0.2

* Limit vertical scrolling to north and south poles

* Limit latitude range to [-80,80]

* Fix some more errors and warnings raised by the upgrade

* Add a release note comment relating to the fix.

* Bump to latest release in 6.0.x track

* Fix one deprecation warning

+++ libraries/cyclestreets-view/src/main/java/net/cyclestreets/views/overlay/LiveItemOverlay.java
@@ -100,7 +100,7 @@ public abstract class LiveItemOverlay<T extends OverlayItem>
   public boolean onZoom(final ZoomEvent event) {
     if (event.getZoomLevel() < zoomLevel_)
       items().clear();
-    zoomLevel_ = event.getZoomLevel();
+    zoomLevel_ = (int)event.getZoomLevel();
     refreshItems();
     return true;
   }
